requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
websocket-client==1.6.1
python-dotenv==1.0.0

# Financial data APIs
alpha-vantage==2.3.1
alpaca-trade-api==3.1.1
finnhub-python==2.4.18

# Technical analysis
ta==0.10.2
yfinance==0.2.18

# Database
sqlite3

# Async support
asyncio
aiohttp==3.8.5

# Logging and monitoring
colorama==0.4.6

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
